// src/app/core/models/suggestion.ts
export class Suggestion {
    SuggestionId: number = null;
    CreateTime: any = null;
    UpdateTime: any = null;
    UpdatedAgent: string = null;
    SentByUserID: number = null;
    SuggestionStatus: string = null;
    Type: string = null;
    FieldName: string = null;
    SuggestedValue: string = null;
    SuggestionComment: string = null;
    details: any;
}